refactor(app): rename v5 upgrade handler to v5.0
