+++ app/upgrades/v5_0/upgrade_test.go
@@ -1,4 +1,4 @@
-package v5_test
+package v5_0_test
 
 import (
 	"testing"
@@ -30,7 +30,7 @@ func (suite *UpgradeTestSuite) TestV5Upgrade() {
 	suite.Setup()
 
 	suite.Ctx = suite.Ctx.WithBlockHeight(dummyUpgradeHeight - 1)
-	plan := upgradetypes.Plan{Name: "v5", Height: dummyUpgradeHeight}
+	plan := upgradetypes.Plan{Name: "v5.0", Height: dummyUpgradeHeight}
 	err := suite.App.UpgradeKeeper.ScheduleUpgrade(suite.Ctx, plan)
 	suite.Require().NoError(err)
 	_, exists := suite.App.UpgradeKeeper.GetUpgradePlan(suite.Ctx)
